billing_api/tests/functional/requirements.txt
pytest==6.1.2
pytest-asyncio==0.12.0
python-dotenv==0.17.0
fastapi==0.63.0
requests==2.25.1
